# nodejs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Unified Data Repository Service API file for subscription data
  description: |
    Unified Data Repository Service (subscription data).
    The API version is defined in 3GPP TS 29.504.
    © 2019, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).
    All rights reserved.
  version: '-'
...
servers:
- url: /
paths:
...
  /subscription-data/{ueId}/ue-update-confirmation-data/sor-data:
    get:
      tags:
      - Authentication SoR (Document)
      summary: Retrieves the SoR acknowledgement information of a UE
      operationId: QueryAuthSoR
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supportedFeatures
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SorData'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AuthenticationSoRDocument
    put:
      tags:
      - Authentication SoR (Document)
      summary: To store the SoR acknowledgement information of a UE
      operationId: CreateAuthenticationSoR
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SorData'
      responses:
        204:
          description: Expected response to a valid request
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AuthenticationSoRDocument
...
  /subscription-data/{ueId}/{servingPlmnId}/provisioned-data/am-data:
    get:
      tags:
      - Access And Mobility Subscription Data (Document)
      summary: Retrieves the access and mobility subscription data of a UE
      operationId: QueryAmData
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: servingPlmnId
        in: path
        description: PLMN ID
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarPlmnId'
      - name: fields
        in: query
        description: attributes to be retrieved
        required: false
        style: form
        explode: false
        schema:
          minItems: 1
          type: array
          items:
            type: string
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      - name: If-None-Match
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.2
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: If-Modified-Since
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.3
        required: false
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: Expected response to a valid request
          headers:
            Cache-Control:
              description: Cache-Control containing max-age, as described in RFC 7234,
                5.2
              style: simple
              explode: false
              schema:
                type: string
            ETag:
              description: Entity Tag, containing a strong validator, as described
                in RFC 7232, 2.3
              style: simple
              explode: false
              schema:
                type: string
            Last-Modified:
              description: Timestamp for last modification of the resource, as described
                in RFC 7232, 2.2
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessAndMobilitySubscriptionData'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AccessAndMobilitySubscriptionDataDocument
...
  /subscription-data/{ueId}/context-data/amf-3gpp-access:
    get:
      tags:
      - AMF 3GPP Access Registration (Document)
      summary: Retrieves the AMF context data of a UE using 3gpp access
      operationId: QueryAmfContext3gpp
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: fields
        in: query
        description: attributes to be retrieved
        required: false
        style: form
        explode: false
        schema:
          minItems: 1
          type: array
          items:
            type: string
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Amf3GppAccessRegistration'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AMF3GPPAccessRegistrationDocument
    put:
      tags:
      - AMF 3GPP Access Registration (Document)
      summary: To store the AMF context data of a UE using 3gpp access in the UDR
      operationId: CreateAmfContext3gpp
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Amf3GppAccessRegistration'
      responses:
        204:
          description: Upon success, an empty response body shall be returned
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AMF3GPPAccessRegistrationDocument
    patch:
      tags:
      - AMF 3GPP Access Registration (Document)
      summary: To modify the AMF context data of a UE using 3gpp access in the UDR
      operationId: AmfContext3gpp
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Expected response to a valid request
        403:
          description: modification is rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AMF3GPPAccessRegistrationDocument
  /subscription-data/{ueId}/context-data/amf-non-3gpp-access:
    get:
      tags:
      - AMF Non-3GPP Access Registration (Document)
      summary: Retrieves the AMF context data of a UE using non-3gpp access
      operationId: QueryAmfContextNon3gpp
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: fields
        in: query
        description: attributes to be retrieved
        required: false
        style: form
        explode: false
        schema:
          minItems: 1
          type: array
          items:
            type: string
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmfNon3GppAccessRegistration'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AMFNon3GPPAccessRegistrationDocument
    put:
      tags:
      - AMF Non-3GPP Access Registration (Document)
      summary: To store the AMF context data of a UE using non-3gpp access in the
        UDR
      operationId: CreateAmfContextNon3gpp
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmfNon3GppAccessRegistration'
      responses:
        204:
          description: Upon success, an empty response body shall be returned
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AMFNon3GPPAccessRegistrationDocument
    patch:
      tags:
      - AMF Non-3GPP Access Registration (Document)
      summary: To modify the AMF context data of a UE using non 3gpp access in the
        UDR
      operationId: AmfContextNon3gpp
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Expected response to a valid request
        403:
          description: modification is rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AMFNon3GPPAccessRegistrationDocument
...
  /subscription-data/{ueId}/context-data/ee-subscriptions/{subsId}/amf-subscriptions:
    get:
      tags:
      - Query AMF Subscription Info (Document)
      summary: Retrieve AMF subscription Info
      operationId: GetAmfSubscriptionInfo
      parameters:
      - name: ueId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                minItems: 1
                type: array
                items:
                  $ref: '#/components/schemas/AmfSubscriptionInfo'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: QueryAMFSubscriptionInfoDocument
    put:
      tags:
      - AMF Subscription Info (Document)
      summary: Create AmfSubscriptions for an individual ee subscriptions of a UE
      operationId: CreateAMFSubscriptions
      parameters:
      - name: ueId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              minItems: 1
              type: array
              items:
                $ref: '#/components/schemas/AmfSubscriptionInfo'
        required: true
      responses:
        204:
          description: Upon success, an empty response body shall be returned
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AMFSubscriptionInfoDocument
    delete:
      tags:
      - Event AMF Subscription Info (Document)
      summary: Deletes AMF Subscription Info for an eeSubscription
      operationId: RemoveAmfSubscriptionsInfo
      parameters:
      - name: ueId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Expected response to a successful subscription removal
      x-swagger-router-controller: EventAMFSubscriptionInfoDocument
    patch:
      tags:
      - AmfSubscriptionInfo (Document)
      summary: modify the AMF Subscription Info
      operationId: ModifyAmfSubscriptionInfo
      parameters:
      - name: ueId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Expected response to a valid request
        403:
          description: modification is rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AmfSubscriptionInfoDocument
...
components:
  schemas:
...
    VarPlmnId:
      pattern: ^[0-9]{5,6}$
      type: string
...
    AccessAndMobilitySubscriptionData:
      type: object
      properties:
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        gpsis:
          type: array
          items:
            $ref: '#/components/schemas/Gpsi'
        internalGroupIds:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/GroupId'
        subscribedUeAmbr:
          $ref: '#/components/schemas/AmbrRm'
        nssai:
          $ref: '#/components/schemas/Nssai'
        ratRestrictions:
          type: array
          items:
            $ref: '#/components/schemas/RatType'
        forbiddenAreas:
          type: array
          items:
            $ref: '#/components/schemas/Area'
        serviceAreaRestriction:
          $ref: '#/components/schemas/ServiceAreaRestriction'
        coreNetworkTypeRestrictions:
          type: array
          items:
            $ref: '#/components/schemas/CoreNetworkType'
        rfspIndex:
          $ref: '#/components/schemas/RfspIndexRm'
        subsRegTimer:
          $ref: '#/components/schemas/DurationSecRm'
        ueUsageType:
          $ref: '#/components/schemas/UeUsageType'
        mpsPriority:
          $ref: '#/components/schemas/MpsPriorityIndicator'
        mcsPriority:
          $ref: '#/components/schemas/McsPriorityIndicator'
        activeTime:
          $ref: '#/components/schemas/DurationSecRm'
        dlPacketCount:
          $ref: '#/components/schemas/DlPacketCount'
        sorInfo:
          $ref: '#/components/schemas/SorInfo'
        upuInfo:
          $ref: '#/components/schemas/UpuInfo'
        micoAllowed:
          $ref: '#/components/schemas/MicoAllowed'
        sharedAmDataIds:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/SharedDataId'
        odbPacketServices:
          $ref: '#/components/schemas/OdbPacketServices'
        subscribedDnnList:
          type: array
          items:
            $ref: '#/components/schemas/Dnn'
...
    Dnn:
      type: string
...
    Amf3GppAccessRegistration:
      required:
      - amfInstanceId
      - deregCallbackUri
      - guami
      - ratType
      type: object
      properties:
        amfInstanceId:
          $ref: '#/components/schemas/NfInstanceId'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        purgeFlag:
          $ref: '#/components/schemas/PurgeFlag'
        pei:
          $ref: '#/components/schemas/Pei'
        imsVoPs:
          $ref: '#/components/schemas/ImsVoPs'
        deregCallbackUri:
          $ref: '#/components/schemas/Uri'
        amfServiceNameDereg:
          $ref: '#/components/schemas/ServiceName'
        pcscfRestorationCallbackUri:
          $ref: '#/components/schemas/Uri'
        amfServiceNamePcscfRest:
          $ref: '#/components/schemas/ServiceName'
        initialRegistrationInd:
          type: boolean
        guami:
          $ref: '#/components/schemas/Guami'
        backupAmfInfo:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/BackupAmfInfo'
        drFlag:
          $ref: '#/components/schemas/DualRegistrationFlag'
        ratType:
          $ref: '#/components/schemas/RatType'
        urrpIndicator:
          type: boolean
        amfEeSubscriptionId:
          type: string
        epsInterworkingInfo:
          $ref: '#/components/schemas/Amf3GppAccessRegistration_epsInterworkingInfo'
    AmfNon3GppAccessRegistration:
      required:
      - amfInstanceId
      - deregCallbackUri
      - guami
      - imsVoPs
      - ratType
      type: object
      properties:
        amfInstanceId:
          $ref: '#/components/schemas/NfInstanceId'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        purgeFlag:
          $ref: '#/components/schemas/PurgeFlag'
        pei:
          $ref: '#/components/schemas/Pei'
        imsVoPs:
          $ref: '#/components/schemas/ImsVoPs'
        deregCallbackUri:
          $ref: '#/components/schemas/Uri'
        amfServiceNameDereg:
          $ref: '#/components/schemas/ServiceName'
        pcscfRestorationCallbackUri:
          $ref: '#/components/schemas/Uri'
        amfServiceNamePcscfRest:
          $ref: '#/components/schemas/ServiceName'
        guami:
          $ref: '#/components/schemas/Guami'
        backupAmfInfo:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/BackupAmfInfo'
        ratType:
          $ref: '#/components/schemas/RatType'
        urrpIndicator:
          type: boolean
        amfEeSubscriptionId:
          type: string
...
    SorData:
      required:
      - provisioningTime
      - ueUpdateStatus
      type: object
      properties:
        provisioningTime:
          $ref: '#/components/schemas/DateTime'
        ueUpdateStatus:
          $ref: '#/components/schemas/UeUpdateStatus'
        sorXmacIue:
          $ref: '#/components/schemas/SorMac'
        sorMacIue:
          $ref: '#/components/schemas/SorMac'
...
    UeUpdateStatus:
      type: string
      enum:
      - NOT_SENT
      - SENT_NO_ACK_REQUIRED
      - WAITING_FOR_ACK
      - ACK_RECEIVED
...
    AmfSubscriptionInfo:
      required:
      - amfInstanceId
      - subscriptionId
      type: object
      properties:
        amfInstanceId:
          $ref: '#/components/schemas/NfInstanceId'
        subscriptionId:
          $ref: '#/components/schemas/Uri'
        subsChangeNotifyCorrelationId:
          type: string
...
    VarUeId:
      pattern: ^(imsi-[0-9]{5,15}|nai-.+|msisdn-[0-9]{5,15}|extid-[^@]+@[^@]+|.+)$
      type: string
    SupportedFeatures:
      pattern: ^[A-Fa-f0-9]*$
      type: string
    ProblemDetails:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/Uri'
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          $ref: '#/components/schemas/Uri'
        cause:
          type: string
        invalidParams:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
    Uri:
      type: string
    InvalidParam:
      required:
      - param
      type: object
      properties:
        param:
          type: string
        reason:
          type: string
    PatchItem:
      required:
      - op
      - path
      type: object
      properties:
        op:
          $ref: '#/components/schemas/PatchOperation'
        path:
          type: string
        from:
          type: string
        value:
          nullable: true
...
    SharedDataId:
      pattern: ^[0-9]{5,6}-.+$
      type: string
...
    Nssai:
      required:
      - defaultSingleNssais
      type: object
      properties:
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        defaultSingleNssais:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Snssai'
        singleNssais:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Snssai'
      nullable: true
    UeUsageType:
      type: integer
    MpsPriorityIndicator:
      type: boolean
    McsPriorityIndicator:
      type: boolean
    DlPacketCount:
      minimum: -1
      type: integer
    SorInfo:
      required:
      - ackInd
      - provisioningTime
      type: object
      properties:
        steeringContainer:
          $ref: '#/components/schemas/SteeringContainer'
        ackInd:
          $ref: '#/components/schemas/AckInd'
        sorMacIausf:
          $ref: '#/components/schemas/SorMac'
        countersor:
          $ref: '#/components/schemas/CounterSor'
        provisioningTime:
          $ref: '#/components/schemas/DateTime'
...
    UpuInfo:
      required:
      - provisioningTime
      - upuAckInd
      - upuDataList
      - upuRegInd
      type: object
      properties:
        upuDataList:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/UpuData_2'
        upuRegInd:
          $ref: '#/components/schemas/UpuRegInd'
        upuAckInd:
          $ref: '#/components/schemas/UpuAckInd'
        upuMacIausf:
          $ref: '#/components/schemas/UpuMac'
        counterUpu:
          $ref: '#/components/schemas/CounterUpu'
        provisioningTime:
          $ref: '#/components/schemas/DateTime'
...
    MicoAllowed:
      type: boolean
...
    NfInstanceId:
      type: string
      format: uuid
...
    GroupId:
      pattern: ^[A-Fa-f0-9]{8}-[0-9]{3}-[0-9]{2,3}-([A-Fa-f0-9][A-Fa-f0-9]){1,10}$
      type: string
...
    Gpsi:
      pattern: ^(msisdn-[0-9]{5,15}|extid-[^@]+@[^@]+|.+)$
      type: string
    AmbrRm:
      required:
      - downlink
      - uplink
      type: object
      properties:
        uplink:
          $ref: '#/components/schemas/BitRate'
        downlink:
          $ref: '#/components/schemas/BitRate'
      nullable: true
    BitRate:
      pattern: ^\d+(\.\d+)? (bps|Kbps|Mbps|Gbps|Tbps)$
      type: string
    RatType:
      anyOf:
      - type: string
        enum:
        - NR
        - EUTRA
        - WLAN
        - VIRTUAL
      - type: string
    Area:
      type: object
      properties:
        tacs:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Tac'
        areaCode:
          $ref: '#/components/schemas/AreaCode'
      oneOf:
      - required:
        - tacs
      - required:
        - areaCode
...
    AreaCode:
      type: string
    ServiceAreaRestriction:
      type: object
      properties:
        restrictionType:
          $ref: '#/components/schemas/RestrictionType'
        areas:
          type: array
          items:
            $ref: '#/components/schemas/Area'
        maxNumOfTAs:
          $ref: '#/components/schemas/Uinteger'
        maxNumOfTAsForNotAllowedAreas:
          $ref: '#/components/schemas/Uinteger'
      allOf:
      - oneOf:
        - not:
            required:
            - restrictionType
        - required:
          - areas
      - anyOf:
        - not:
            required:
            - restrictionType
            properties:
              restrictionType:
                type: string
                enum:
                - NOT_ALLOWED_AREAS
        - not:
            required:
            - maxNumOfTAs
      - anyOf:
        - not:
            required:
            - restrictionType
            properties:
              restrictionType:
                type: string
                enum:
                - ALLOWED_AREAS
        - not:
            required:
            - maxNumOfTAsForNotAllowedAreas
...
    CoreNetworkType:
      anyOf:
      - type: string
        enum:
        - 5GC
        - EPC
      - type: string
    RfspIndexRm:
      maximum: 256
      minimum: 1
      type: integer
      nullable: true
    DurationSecRm:
      type: integer
      nullable: true
    OdbPacketServices:
      nullable: true
      anyOf:
      - type: string
        enum:
        - ALL_PACKET_SERVICES
        - ROAMER_ACCESS_HPLMN_AP
        - ROAMER_ACCESS_VPLMN_AP
      - type: string
...
    PurgeFlag:
      type: boolean
    ImsVoPs:
      anyOf:
      - type: string
        enum:
        - HOMOGENEOUS_SUPPORT
        - HOMOGENEOUS_NON_SUPPORT
        - NON_HOMOGENEOUS_OR_UNKNOWN
      - type: string
    DualRegistrationFlag:
      type: boolean
...
    DateTime:
      type: string
      format: date-time
...
    SorMac:
      pattern: ^[A-Fa-f0-9]{32}$
      type: string
...
    AckInd:
      type: boolean
...
    Pei:
      pattern: ^(imei-[0-9]{15}|imeisv-[0-9]{16}|.+)$
      type: string
...
    Guami:
      required:
      - amfId
      - plmnId
      type: object
      properties:
        plmnId:
          $ref: '#/components/schemas/PlmnId'
        amfId:
          $ref: '#/components/schemas/AmfId'
...
    BackupAmfInfo:
      required:
      - backupAmf
      type: object
      properties:
        backupAmf:
          $ref: '#/components/schemas/AmfName'
        guamiList:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Guami'
...